Split the two graphs in Graphs.tsx into separate files
It's hard to follow the code flow when Graphs.tsx is this large. I think
splitting the Power bar and chart graphs into separate component will
help with legibility.
