+++ apps/frontend/src/components/Graphs.tsx
@@ -2,21 +2,11 @@ import { AspectRatio, Grid, Stack } from '@chakra-ui/layout';
 import { useBreakpointValue } from '@chakra-ui/media-query';
 import { Member } from '@dundring/types';
 import * as React from 'react';
-import {
-  Area,
-  AreaChart,
-  Bar,
-  BarChart,
-  ResponsiveContainer,
-  Tooltip,
-  YAxis,
-} from 'recharts';
-import { hrColors, powerColors, untrackedColor } from '../colors';
 import { useData } from '../context/DataContext';
 import { LocalRoom } from '../context/WebsocketContext';
-import { CustomChartTooltip } from './Graph/CustomChartTooltip';
-import { CustomGraphTooltip } from './Graph/CustomGraphTooltip';
 import { ShowData } from './Graph/GraphContainer';
+import { PowerBar } from './PowerBar';
+import { DataPoint, PowerChart } from './PowerChart';
 
 interface Props {
   otherUsers: Member[];
@@ -25,10 +15,14 @@ interface Props {
   showUserData: ShowData;
   showOtherUsersData: { [username: string]: ShowData };
 }
-const mergeArrays = (arr1: any[], arr2: any[]) => {
+
+const mergeArrays = <T1 extends object, T2 extends object>(
+  arr1: T1[],
+  arr2: T2[]
+) => {
   const a = arr1.length > arr2.length ? arr1 : arr2;
   const b = arr1.length > arr2.length ? arr2 : arr1;
-  return a.map((a, i) => ({ ...a, ...b[i] }));
+  return a.map((aItem, i) => ({ ...aItem, ...b[i] }));
 };
 
 export const Graphs = ({
@@ -41,7 +35,8 @@ export const Graphs = ({
   const { data: laps, untrackedData: rawUntrackedData } = useData();
   const rawData = laps.flatMap((x) => x.dataPoints);
   const numPoints = 500;
-  const [allMerged, myAvgPower, otherPeoplesAvgPower] = React.useMemo(() => {
+
+  const allMerged = React.useMemo(() => {
     const data = rawData.map((dataPoint) => ({
       'You HR': dataPoint.heartRate,
       'You Power': dataPoint.power,
@@ -56,7 +51,7 @@ export const Graphs = ({
       .map((user) => {
         const data = activeGroupSession?.workoutData[user.username];
         if (!data) return [];
-        const baseData = [
+        const baseData: DataPoint[] = [
           ...new Array(numPoints).fill({
             [user.username + ' HR']: undefined,
             [user.username + ' Power']: undefined,
@@ -75,229 +70,71 @@ export const Graphs = ({
       })
       .reduce(
         (merged, data) => mergeArrays(merged, data),
-        [...Array(numPoints)]
+        [...(Array(numPoints) as DataPoint[])]
       );
 
     const filledData = [
-      ...new Array(numPoints).fill({
+      ...(new Array(numPoints).fill({
         'You HR': undefined,
         'You Power': undefined,
-      }),
+      }) as DataPoint<'You'>[]),
       ...data,
     ].splice(-numPoints);
 
     const filledUntrackedData = [
-      ...new Array(numPoints).fill({
+      ...(new Array(numPoints).fill({
         'You-Untracked HR': undefined,
         'You-Untracked Power': undefined,
-      }),
+      }) as DataPoint<'You-Untracked'>[]),
       ...untrackedData,
     ].splice(-numPoints);
 
-    const otherPeoplesAvgPower = otherUsers.map((user) => {
-      const data = activeGroupSession?.workoutData[user.username];
-      if (!data) return null;
-      const avgPower = Math.floor(
-        ((data[0]?.power || 0) +
-          (data[1]?.power || 0) +
-          (data[2]?.power || 0)) /
-          3
-      );
-      const name = `${user.username} Power`;
-      return { name, [name]: avgPower };
-    });
-    const myAvgPower = Math.floor(
-      [...rawData]
-        .splice(-3)
-        .reduce((sum, data) => sum + (data.power || 0), 0) / 3
+    return mergeArrays(
+      mergeArrays(filledData, otherPeoplesDataMerged),
+      filledUntrackedData
     );
+  }, [rawData, rawUntrackedData, otherUsers, activeGroupSession]);
 
-    return [
-      mergeArrays(
-        mergeArrays(filledData, otherPeoplesDataMerged),
-        filledUntrackedData
-      ),
-      myAvgPower,
-      otherPeoplesAvgPower,
-    ];
-  }, [activeGroupSession?.workoutData, otherUsers, rawData, rawUntrackedData]);
+  const myAvgPower = Math.floor(
+    [...rawData].splice(-3).reduce((sum, data) => sum + (data.power || 0), 0) /
+      3
+  );
 
-  const fillAreaChart = (
-    dataPrefix: string,
-    checked: ShowData,
-    index: number,
-    showFill: boolean,
-    type: 'hr' | 'power',
-    isUntracked: boolean = false
-  ) => {
-    const hrColor = isUntracked
-      ? untrackedColor
-      : hrColors[index % hrColors.length];
-    const powerColor = isUntracked
-      ? untrackedColor
-      : powerColors[index % powerColors.length];
-    const hrGradientId = dataPrefix + 'colorHR';
-    const powerGradientId = dataPrefix + 'colorPower';
-    const showHr = type === 'hr' && checked.hr;
-    const showPower = type === 'power' && checked.power;
-    return (
-      <React.Fragment key={index}>
-        {showFill ? (
-          <defs>
-            {showHr ? (
-              <linearGradient
-                id={`${hrGradientId}`}
-                x1="0"
-                y1="0"
-                x2="0"
-                y2="1"
-              >
-                <stop offset="5%" stopColor={hrColor} stopOpacity={0.4} />
-                <stop offset="95%" stopColor={hrColor} stopOpacity={0} />
-              </linearGradient>
-            ) : null}
-            {showPower ? (
-              <linearGradient
-                id={`${powerGradientId}`}
-                x1="0"
-                y1="0"
-                x2="0"
-                y2="1"
-              >
-                <stop offset="5%" stopColor={powerColor} stopOpacity={0.4} />
-                <stop offset="95%" stopColor={powerColor} stopOpacity={0} />
-              </linearGradient>
-            ) : null}
-          </defs>
-        ) : null}
-
-        {showHr ? (
-          <Area
-            dataKey={`${dataPrefix} HR`}
-            stroke={hrColor}
-            fill={`url(#${hrGradientId})`}
-            animationDuration={0}
-          />
-        ) : null}
-        {showPower ? (
-          <Area
-            dataKey={`${dataPrefix} Power`}
-            stroke={powerColor}
-            fill={`url(#${powerGradientId})`}
-            animationDuration={0}
-          />
-        ) : null}
-      </React.Fragment>
+  const otherPeoplesAvgPower = otherUsers.map((user) => {
+    const data = activeGroupSession?.workoutData[user.username];
+    if (!data) return null;
+    const avgPower = Math.floor(
+      ((data[0]?.power || 0) + (data[1]?.power || 0) + (data[2]?.power || 0)) /
+        3
     );
-  };
-
-  const fillBarChart = (dataPrefix: string, show: boolean, index: number) => {
-    const powerColor = powerColors[index % powerColors.length];
-    const powerGradientId = dataPrefix + 'colorPower';
-    return (
-      <React.Fragment key={index}>
-        {showFill ? (
-          <defs>
-            <linearGradient
-              id={`${powerGradientId}`}
-              x1="0"
-              y1="0"
-              x2="0"
-              y2="1"
-            >
-              <stop offset="5%" stopColor={powerColor} stopOpacity={0.8} />
-              <stop offset="95%" stopColor={powerColor} stopOpacity={0} />
-            </linearGradient>
-          </defs>
-        ) : null}
-
-        {show ? (
-          <Bar
-            dataKey={`${dataPrefix} Power`}
-            stroke={powerColor}
-            fill={`url(#${powerGradientId})`}
-          />
-        ) : null}
-      </React.Fragment>
-    );
-  };
+    const name = `${user.username} Power`;
+    return { name, [name]: avgPower };
+  });
 
   const showPowerBar = useBreakpointValue({ base: false, md: true });
+
   return (
     <Stack width="100%">
       <Grid templateColumns={showPowerBar ? '5fr 1fr' : '1fr'}>
         <AspectRatio ratio={16 / 9} width="100%">
-          <ResponsiveContainer>
-            <AreaChart data={allMerged}>
-              {otherUsers.map((user, i) =>
-                fillAreaChart(
-                  user.username,
-                  showOtherUsersData[user.username] || {
-                    hr: true,
-                    power: true,
-                  },
-                  i + 1,
-                  showFill,
-                  'power'
-                )
-              )}
-              {fillAreaChart(
-                'You-Untracked',
-                showUserData,
-                0,
-                showFill,
-                'power',
-                true
-              )}
-              {fillAreaChart('You', showUserData, 0, showFill, 'power')}
-              {otherUsers.map((user, i) =>
-                fillAreaChart(
-                  user.username,
-                  showOtherUsersData[user.username] || {
-                    hr: true,
-                    power: true,
-                  },
-                  i + 1,
-                  showFill,
-                  'hr'
-                )
-              )}
-              {fillAreaChart('You', showUserData, 0, showFill, 'hr')}
-              {fillAreaChart(
-                'You-Untracked',
-                showUserData,
-                0,
-                showFill,
-                'hr',
-                true
-              )}
-
-              <YAxis />
-              <Tooltip content={<CustomGraphTooltip />} />
-            </AreaChart>
-          </ResponsiveContainer>
+          <PowerChart
+            allMerged={allMerged}
+            otherUsers={otherUsers}
+            showOtherUsersData={showOtherUsersData}
+            showUserData={showUserData}
+            showFill={showFill}
+          />
         </AspectRatio>
-        {showPowerBar ? (
-          <AspectRatio ratio={1} width="100%">
-            <ResponsiveContainer>
-              <BarChart
-                data={[
-                  [
-                    { name: 'You Power', 'You Power': myAvgPower },
-                    ...otherPeoplesAvgPower,
-                  ].reduce((res, cur) => ({ ...res, ...cur }), {}),
-                ]}
-              >
-                {fillBarChart('You', true, 0)}
-                {otherUsers.map((user, i) =>
-                  fillBarChart(user.username, true, i + 1)
-                )}
-                <YAxis />
-                <Tooltip content={<CustomChartTooltip />} cursor={false} />
-              </BarChart>
-            </ResponsiveContainer>
+        {showPowerBar && (
+          <AspectRatio ratio={1}>
+            <PowerBar
+              showFill={showFill}
+              otherUsers={otherUsers}
+              otherPeoplesAvgPower={otherPeoplesAvgPower}
+              myAvgPower={myAvgPower}
+            />
           </AspectRatio>
-        ) : null}
+        )}
       </Grid>
     </Stack>
   );
